Add os-specific instructions for linking settings

+++ link_configs.sh
@@ -1,8 +1,21 @@
-#!/bin/sh
+#!/bin/bash
 
-mkdir -p $HOME/.config/nvim
-ln -s $HOME/.dotfiles/init.vim $HOME/.config/nvim/init.vim
-ln -s $HOME/.dotfiles/.tmux.conf $HOME/.tmux.conf
-ln -s $HOME/.dotfiles/.spacemacs $HOME/.spacemacs
+if [[ ! -f $HOME/.config/nvim/init.vim ]]; then
+  mkdir -p $HOME/.config/nvim
+  ln -s $HOME/.dotfiles/init.vim $HOME/.config/nvim/init.vim
+fi
 
-echo ". ~/.dotfiles/.zshrc" > ~/.zshrc
+[[ -f $HOME/.tmux.conf ]] || ln -s $HOME/.dotfiles/.tmux.conf $HOME/.tmux.conf
+[[ -f $HOME/.spacemacs ]] || ln -s $HOME/.dotfiles/.spacemacs $HOME/.spacemacs
+[[ -f $HOME/.zshrc ]] || echo ". ~/.dotfiles/.zshrc" > ~/.zshrc
+
+os="`uname`"
+if [[ "$os" == 'Darwin' ]]; then
+  mkdir -p "$HOME/Library/Application Support/Code/User"
+  ln -sf $HOME/.dotfiles/Code/User/settings.json "$HOME/Library/Application Support/Code/User/settings.json"
+  ln -sf $HOME/.dotfiles/Code/User/keybindings.json "$HOME/Library/Application Support/Code/User/keybindings.json"
+elif [[ "$os" == 'Linux' ]]; then
+  mkdir -p $HOME/.config/Code/User/
+  ln -sf $HOME/.dotfiles/Code/User/settings.json $HOME/.config/Code/User/settings.json
+  ln -sf $HOME/.dotfiles/Code/User/keybindings.json $HOME/.config/Code/User/keybindings.json
+fi
